python-backend/requirements.txt
fastapi>=0.111.1,<1.0.0
uvicorn>=0.30.4,<1.0.0
requests>=2.32.3,<3.0.0
databricks-sdk>=0.29.0,<1.0.0
cryptography>=43.0.0,<44.0.0
python-dotenv>=1.0.1,<2.0.0